refact: split parse_error.rs
