+++ src/parser/error/mod.rs
@@ -1,3 +1,5 @@
+mod error_code;
 mod parse_error;
 
-pub use parse_error::{ErrorCode, ParseError};
+pub use error_code::ErrorCode;
+pub use parse_error::ParseError;
